backend/requirements.txt
# ADUmedia Website Backend Dependencies

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
supabase==2.12.0

# Authentication
PyJWT==2.10.1
bcrypt==4.2.1

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# HTTP client (for webhooks)
httpx==0.28.1
aiohttp==3.11.11

# Date/time
python-dateutil==2.9.0.post0

# Environment
python-dotenv==1.0.1

# CORS
starlette==0.45.3
